Type singletons (#9)
This patch set makes the ast type system act more like LLVM's internally, 
namely, pointer equality comparison is enough to check if two types are equal.

* Start comparing types only by pointer

Currently, type comparisions take a virtual function call, as one "type"
may have multiple instances.
LLVM uses pointer comparisions, which are more performant. It can ensure
that a type has one unique instance, as all types are factory created.

Note that tests are failing, as the factory system has not been
implemented yet.

* Remove `operator==` from ast::type

Types are now only pointer compared.

* Make all ast types non_movable

This is a step towards making the types singletons.

* Make `pointed_to` private

This prevents accidental mutation, which is needed for the type
singleton system.

* Private the user_type and function_type data

* Pointer types are created from factory functions

* `create` for the nullable and nonnullable pointer types do uniqueness

* Make user types unique

* Make function types unique

Note that because function types are parameterized over return type and
args, it is currently easier to do a linear search than to use a map.