Structs (#10)
Additional changes:

- Format cmake files
- Mandate then in if-else
- Add a scoped_map utility

* Prototype grammar for structs

The only issue with the way structs are so far defined is
```
if identifier {
```
The identifier could either be
1. the type name of a struct (in which case we need to shift the `{`), or
2. the name of a variable (we need to reduce `identifier`)

* Fixed all conflicts

Now, lets require semicolons and ifs need either parens or `then`.

* `if-then` will be required in all cases

This disambiguates between `if foo {` where `foo` is either a typename
followed by a struct initalization or a variable followed by a compound
statement.

* Mandate `then` after `if condition`

This disambiguates the if condition from the if body. This is needed for
cases such as `if foo {`, as when structs will be added, the `foo` could
either be
1. a variable (and thus the condition) with the `{` being the start of a
   compound statement, or
2. a typename with the `{` being the start of a struct initialization

* Let statements require semicolons

This is to disambiguate between a let of a struct
```
let x = typename {
	// initialization
};
```
and a let followed by a compound statement
```
let x = foo;
{
}
```

* Parsed struct declarations

* Parsed struct initialization

* Formatting pass

* Add a member access operand for binary_expr

* Removed accidental include in global_map.hpp

* Add accessor for the module_name

* Add ast type support for struct types

* Add visitor support for struct declarations

* Print out struct_type instead of user_type

* Lower struct declarations into LLVM

* Actually check that `id {` and `id (` are different

* Lex the dot operator

* Parse member access

* Parse if expressions as atoms

This allows `1 + (if x then 2 else y + 2)` to be written without the
parenthesis.

* Formatting pass

* Update to LLVM 14

TargetRegistry.h was moved and `getElementType` has been deprecated in
favor of `getPointerElementType`.

* Add `noexcept` where it should have been

* Use the CMake package method of loading LLVM

There are many linker errors when trying to get all targets, so to
simplify the process, support was removed until a fix can be found.

* Implement a scoped_map template for types and values

The pattern of a `std::vector<std::map<K, V>>` to represent a stack of
scopes should be explicitly named as such.

* Formatting pass with new CMake format

* Register user declared structs in the type checker

* Added struct_init to the `expand_node_macro`

* `global_map` prevents empty string as a key.

* Patch pernicious bug regarding inappropriate `std::move`

* Type check struct declarations and member accesses.

* Add `true` and `false` as boolean literals

* Rename `lexer::module_name()` to `file_name()`

* The parser now knows what module it is parsing

* Permit struct types to be reused

* Type check struct initializers

* Add debug printing for `global_map`

* Implement codegen for structure member access

* Implement codegen for struct initializations

* `type_context` now uses `shared_ptr<ast::type>`

This ensure that types are unique per module.

* Structs need to be passed as pointers

This is fine until we allow mutation.