Allow the cfg builder to know when a node comes from an id lookup
This is used when chaining nodes for function calls, as they are
normally evaluated left-to-right, but this leads to issues when an
argument is just an identifier.

The solution is to ignore identifiers for flow chaining, but we do not
know if something is from identifier lookup at the use site.