Fix constants in if expressions failing LLVM lowering
Consider `if cond then some expr else constant`. We do not generate a
new block for the constant as that would be wasteful (probably a
premature optimization). This used to confuse the lowering code,
as the constant needs some source block, but we point to the phi block.

The solution is currently to find a block that calls us a successor and
use that as the source.